requirements.txt
mpi4py>=3.1.3
opencv-python>=4.5.3
opencv-contrib-python>=4.5.3
numpy>=1.20.0
pillow>=8.3.1
pyyaml>=6.0
matplotlib>=3.4.3
cuda-python>=11.6.0
torch>=1.10.0
requests>=2.28.0 
pandas>=1.3.0
ultralytics>=8.0.0
